refactor: 폴더 구조를 더 명확하게 재조정
