+++ src/assets/icons/SearchIcon.js
@@ -1,0 +1,20 @@
+import React from 'react';
+
+const SearchIcon = () => {
+    return (
+        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16" fill="none">
+            <path d="M7.84442 13.8363C11.1539 13.8363 13.8368 11.1534 13.8368 7.84393C13.8368 4.53444 11.1539 1.85156 7.84442 1.85156C4.53493 1.85156 1.85205 4.53444 1.85205 7.84393C1.85205 11.1534 4.53493 13.8363 7.84442 13.8363Z" 
+                  stroke="#666666" 
+                  strokeWidth="1.5" 
+                  strokeLinecap="round" 
+                  strokeLinejoin="round"/>
+            <path d="M12.0122 12.3232L14.3616 14.6665" 
+                  stroke="#666666" 
+                  strokeWidth="1.5" 
+                  strokeLinecap="round" 
+                  strokeLinejoin="round"/>
+        </svg>
+    );
+}
+
+export default SearchIcon;
