+++ src/assets/icons/YoutubeIcon.js
